requirements.txt
# Core
Django==5.1.4
python-dotenv==1.2.1

# Database
psycopg[binary]==3.3.2

# Authentication
django-allauth==65.13.1
requests==2.32.3

# Caching
django-redis==6.0.0
redis==7.1.0

# Media files
Pillow==12.0.0

# Utils
python-slugify==8.0.4

gunicorn==21.2.0

# API
djangorestframework==3.15.2

django-filter==24.3

djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.2

